+++ src/error.rs
@@ -1,29 +1,10 @@
-// qr: Encode URLs or text into QR codes.
-// Copyright (C) 2022 Marco Radocchia, 2024 pepa65
-//
-// This program is free software: you can redistribute it and/or modify it under
-// the terms of the GNU General Public License as published by the Free Software
-// Foundation, either version 3 of the License, or (at your option) any later
-// version.
-//
-// This program is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-// FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
-// details.
-//
-// You should have received a copy of the GNU General Public License along with
-// this program. If not, see https://www.gnu.org/licenses/.
+// qr - Encode text into svg/png/jpg/terminal format QR codes
 
 use std::{
     fmt::{self, Display, Formatter},
     io::{self, Write},
 };
 use termcolor::{BufferWriter, Color, ColorChoice, ColorSpec, WriteColor};
-
-pub enum Warning {
-    /// Occurs when scale argument is provided with non-raster image output format.
-    UnexpectedScaleOpt,
-}
 
 pub enum Error {
     /// Occurs when unable to generate QR code.
@@ -40,7 +21,6 @@ pub enum Error {
 }
 
 pub enum ErrorKind {
-    Warning(Warning),
     Error(Error),
 }
 
@@ -56,7 +36,6 @@ impl ErrorKind {
         //  * Warning -> ("warning:", Yellow)
         //  * Error -> ("error:", Red)
         let (prefix, color) = match self {
-            Self::Warning(_) => ("warning", Some(Color::Yellow)),
             Self::Error(_) => ("error", Some(Color::Red)),
         };
 
@@ -76,12 +55,6 @@ impl Display for ErrorKind {
     /// Print colored error message, but ONLY on Stderr stream.
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Warning(warning) => match warning {
-                Warning::UnexpectedScaleOpt => write!(
-                    f,
-                    "ignoring `scale` CLI option with non-raster output format"
-                ),
-            },
             Self::Error(error) => match error {
                 Error::QrCodeErr(msg) => write!(f, "unable to generate QR code: {msg}"),
                 Error::InvalidOutputExt => write!(f, "invalid output file extension"),